add SimpleBatchMint tool
This new tool will allow for users to select the ARC3/19/69, enter the IPFS CID for the collection, Name with option for starting #, Unit Name with option for starting #, have toggles for freeze and CB, Description, External URL, Creator, Token, and Royalty. They can then upload a sheet with just the traits to mint fast and easy.