Improve scrolling to top for longer pages
For some longer pages scrolling to the top of the header was not enough for some reason. So now we use the root as a reference instead of the header. Should also fixes scroll issues for instances where there's an attempt to scroll to the top of a header that has not been rendered yet.